separate getGIVpowerBalanceByAddress
